text_extraction/requirement.txt.txt
pytorch
flask
numpy
pandas 
transformers
tokenizers
